validate: Validate XML but never fail during that stage
The idea here is that we always want all parts of the `daps validate`
to run. The DocBook XML validation is the part that tends to take
longest, so it really should run as part of the target.

Pretty much everything else does not take too long and can thus run
before (as was already set up in 6ee4e644 "Fine tuning validation).

The leading `-` makes GNU Make not fail on non-0 exit codes. We still
need to store the exit code and we do that in a file `.xml_valid`. That
is extremely ugly though--I wish I could just set a Make variable
somehow.